feat: renamed wrappers
